chainpack: fix packing of empty string
Strings are packed with size in front and that requires direct call to
the write_uint_data and that uses in proxy
_significant_bits_part_length. The issue here is that for zero this
method says that bits count is also 0 and thus causes error "bytearray
index out of range". This ensures that we always return at least one.

This also modifies names of the variables in the function to match
Python conventions.