Suypport standard SHV error repesentation
This adds shverror and set_shverror to the RpcMessage class and also
Enum with SHV error codes. They implement the correct error
representation for SHV instead of just plain access to the error field
of the original error and set_error. The change is intentionally
compatible to ensure that we won't break existing code.