Add ShvServer and rework ShvClient and ClientConnection
The ShvServer provides a simple implementation of the server that waits
for clients to be connected. The arguments are intentionally consistent
with those in ShvClient. Server initializes ShvClient for new
connections and passes it to the callback.

The rework of ShvClient consists of change in initialization and
connection method. ShvClient now has to be initialized with writter and
reader. The connection method is now class method and initializes
ShvClient for you. This also results in removal of connection state as
we no longer have pretty much anything to track because existing object
can be either connected (including the login when accessible by the
user when connect is used) or disconnected. To detect disconnect you
should use reader and check for EOF.