rpcclient: make login required
The idea that we can skip login was to skip unnecessary step when
connecting on already verified connection such as local socket. The
issue is that we do not use it only for login but also to pass options
to the broker and this should be possible to do even on local socket.
The idea is to instead just pass `None` as user name and password and
not check that when client connects over local socket.