Use native Python types instead of custom RpcValue when ever possible
WARNING: This is breaking change that requires a lot of small code
changes in the applications using this!

The motivation behind this is to remove boilerplate that was created by
RpcValue and facilitate rather native Python type checking over checking
based on enum.

The important effect is that values are returned directly and thus
indexing list inside list is now `lst[x][y]` instead of previous
`lst.value[x].value[y]`.

The `isinstance` should be used now to detect type of the value.

The meta is facilitated trough `SHVMeta` class and child of it. Most
likely you want to use `SHVMeta.new` to get an appropriate meta capable
class. Those classes are designed to be compatible and in most cases
based on the native types.

Not every type could have been directly mapped to the Python variant and
thus there are these exceptions:
* `SHVNull` does not relate to `None` and thus you have to check for it
  explicitly alongside with it. `is_shvnull` is provided for this purpose.
* `SHVBool` does not relate to `bool` and thus you have to check for it
  explicitly alongside with it. `is_shvbool` is provided for this purpose.
* `SHVDict` is used for both Map and IMap. `is_shvmap` and `is_shvimap`
  are provided to detect these based on the content. Note that empty
  dictionary matches both Map and IMap check.