Add SimpleClient and remove ClientConnection
SimpleClient is now common base utility for connecting to the SHV
Broker. It provides minimal API to handle requests and signals and to
perform calls on its own.

SimpleClient mostly replaces ClientConnection with exception of signal
handling through change notifications. To provide the same functionality
an extension to the SimpleClient named ValueClient is provided. With it
you can register callback on value change as well as to load snapshot
from logs.