Add RpcUrl
The connection to the broker requires considerable amount of parameters.
This is common for the connection and there is also a common solution
for this in form of URL.

The URL as described here is not any standard SHV RPC URL. It is rather
just an initial attempt to have something like that.