simpleclient: merge DeviceClient to SimpleClient
We want to support client identification (appName and appVersion) for
all clients and easiest implementation was already in DeviceClient.

This also changes `_ls` and `_dir` in such a way they are generators not
returning lists because that way we do not have to assemble multiple
lists just to incrementally add some values.

RpcMethodDesc was added to make working with method description more
systematic. `dir_details` method now returns list of it.

This commit also moves ValueClient to the separate file.