broker: provide a way to trust granted access levels on su level
This provides a way to trust to access level that was in the message
(not overwriting it user's one). The use case is for broker chaining
where access level deduced by first broker is then passed around to the
device. The su level is used for this feature because it is the highest
access level and when client is allowed to specify granted access level
it can specify su and thus have highest access level anyway.

Tests now use development access level to bypass this feature in regular
communication testing.