chainpack: fix decimal chainpack packing
The decimal method is_integer_ration returns ratio really and it might
not be in decimal form (such as 12E-3 is returned as 3/250).

The replacement is to use as_tuple method that returns mantisa and
exponent. The only issue is that mantisa is returned as tuple of digits
which needs to be converted to integer.