rpcurl: allow login on local socket and change host to location
This contains two changes. Let's first talk about incompatibility in the
code we introduced:
The attribute of RpcUrl that was previously named `host` is now named
`location`. The original name was taken from asyncio API but that was
only for TCP. We also support local sockets and having them pass their
path in attribute named host is really confusing.

The second change is just a new feature wrapped as change. The parsing
of login options is now done unconditionally and thus we can perform
login on local socket now.