cpon: correctly decode and encode c-strings
The previous code took data byte by byte but that way it interpreted
some multi-bytes characters as a different ones. To correctly perform
decoding we should rather just use Python native way.

The escape decoding on its own was also invalidly implemented and all
characters were decode as '\'.

It also used the same method to escape string as blob. That resulted in
invalid escapes to be used in the packed string.