gitlab-ci: add devenv stage
Steps in the pipeline are using Nix development environment and because
some of them run in parallel we end up with them building the same stuff
in parallel. This instead builds development environment initially and
thus subsequent stages can use cache and thus we speed up the whole
pipeline and reduce the duplicate work.

This also fixes issue where local builds were cached but not used
because they were untrusted. We need to sign them and for that we can
generate cache specific signing key that we just store as part of the
cache.