simpleclient: fix ping sending on listening connections
The issue here was that we updated inactivity timer when we received
some data. This works for broker but for clients connected to the broker
this can result in ping not sent due to constant receive of message from
the broker.

There needs to be a two activity timers, one for receive and one for
send because broker clients are interested in time of last receive while
clients connected to the broker need to know hen they sent the last
message.