simpleclient: support multiple call attempts
Any message on SHV can be lost and thus when method is being called
either request or response message can be lost and thus call would not
finish. This provides easy to use functionality that can try call
multiple times and in the end fail with exception if call just wasn't
working.