simpleclient: allow more dynamic ping behavior
The common approach is that broker disconnects idle sessions but thanks
to the `.app:ping` being required for every single client this might not
be the only way to construct the SHV network. To keep doors open
SimpleClient now has this behavior rather selectable to reduce need to
duplicate disconnect code when that is needed instead of ping.