Rework version specification
The previous way was problematic because it was hard to correctly
receive the version of the package. The common problem is when running
the local version but version from system would be reported.

This is in the end more complex way over just having it in
pyproject.toml because it must be read from file in every single tool
but in the end it allows systematic version tracking.