broker: Minimize access level instead of overwrite
This allows better broker chaining. It is not standard behavior right
now but pySHV is a good test bench to verify behavior.

This also adds experimental Access Level field that right now is kept in
sync with field in Access. In the future it might be possible to slowly
migrate to that level if this solution gets validated.