value: improve type checking
This allows more robust type checking of Python types. It performs deep
inspection but won't tell you the location of the wrong type but still
that is a lot more versatile than having to write these checks manually.

The change here is also that is_shvmap and is_shvimap are now also
checking their values not just keys.