broker: filter signal access based on the source method
The broker's configuration specifies accessible methods. The original
design was that signals are unrelated to the methods and in such case
filtering by signal name made sense but now all signals must be
associated with some method so we should rather filter access based on
the access level to that method.

It is questionable if filtering based on the signal name is usable and
thus it is not added at the moment as there would have to be an
additional configuration option in the broker to set it.