tests: Do not depend on libshv
This makes pyshv a standalone tool. It is self checking.

The issue might be that now SimpleClient tests depend on Broker that
depends on SimpleClient to work in most cases. But all is not that bad
because core of the Broker is implemented on top of the SimpleBase that
is not tested separately anyway. This applies only to the simple
abstractions. The message exchanging is well tested without it and thus
we can rely on that. It just might not be directly visible if bug is
introduced in SimpleBase that it is in there.