shv/broker: rework to support TTL in subscriptions and autosetup
This completely reworks how RpcBroker works. The primary concept change
is the way clients are managed. Now they are tracked by RPC servers and
not by broker and thus broker tracks only those clients that are active
and clients register and unregister themself based on their state.

The message propagation in Broker also changes to instead using queue.
This prevents from broker block when one of the peers gets stuck.