rpcri: remove RpcRI and use it only through associated functions
In the end we do not have to split RI to its componenets in most cases
and be creating object just to hold them split is kind of stupid. This
instead only introduces few functions that are intended to be used with
RPC RI strings and RIs them self are kept as strings.