shv/broker: Generate nonce only once per session
This is required because on some transport layers it is possible that
clients send multiple hello requests that are buffered before we process
them. This then causes fast rollover of nonce but client will react on
the first one and thus send in the end invalid login (with at that time
no longer valid nonce).