treewide: rename Simple* API to SHV*
The "Simple" in the name of this API is misleading especially if
included in the bigger project. This instead uses more correct SHV
prefix. The usage is not changed except that name is now more
appropriate.

This is incompatible API change but code modifications should be really
easy in the downstream projects and thus must happen so why not now.