rpcfile: add utility to use and implement RPC files
This is support for both implementation of the RPC file as well as class
to be used to access the RPC file.

There is still missing implementation of RpcTextFile that would be based
on top of the RpcFile and provide the textual access. The RpcFile
buffered access is also not that tested and in some combinations of
actions might be problematic. It is just an initial version.