Add shvtypes utility to work with RPC type hints
The newly introduced RPC type hints is the definition for the parameter
and result of the methods description. This provides a way to send more
compact data while retrieving additional representation info from the
type hint provided in the 'dir' result.

The two use cases are expected at the moment:
* The nicer printing of complex map data and units for integers.
* The possibility to validate parameter even before it is sent as well
  as a more informative parameter input (such as table for structures)

To support that the object representation can be inspected and defined
validation method can be also used. Note that validation doesn't inform
why value is invalid only that it is. This might be added in the future
either as a dedicated API or as an API breaking change (the result would
most likely have to be string to provide issue description).