value: remove confusion between Map and IMap for empty mappings
The empty mappings can be both Map as well as IMap. Because our packers
check for IMap first the empty Map is packed as IMap instead. This
ensures that functions `is_shvmap` doesn't accept `SHVIMap` objects and
vice versa for the `is_shvimap`. The unpackers now also convert empty
mappings to the `SHVMap` and `SHVIMap` classes to explicitly encode the
unpacked type.