release.sh: fix invalid twine check before we do release and the release
The `python3 -m build` actually generates files to `dist` directory.

The syntax `${0%/*}` removes tail and thus expectadly name of the script
but if there is no slash then name is kept. This happens pretty much
only if script it launched by `bash release.sh` but that is exactly what
we do in the Gitlab CI.