rename.sh: improve the experience
This changes the rename script to give more hints about what is to be
replaced by which input. This also changes order of the taken input to
me more in line with common naming approach. The first input is the full
name of the project as stated in the readme. The second input is the
name of the python project that must be compatible with python project
naming scheme and lastly this name is used or custom input is taken to
name the python package.