rpctransport/tty: Fix TypeError when TTY client is closed
The executor that reads the serial port is not terminated gracefully by
aioserial when close is called and thus it touches some variables set to
`Note` at that point causing `TypeError` to be raised. We handle this as
disconnect because it should happen only on disconnect.

The fix seems to be present in PR
https://github.com/mrjohannchang/aioserial.py/pull/7 but it seems that
development of this project somewhat stalled. We should think about
moving to some alternative if nothing happens in year or two.