rpcerrors: remove no longer standard errors
These errors were removed from SHV standard because they were originally
only intended for the internal use in libshv (original C++
implementation) or are no longer relevant. For this reason they should
not be provided in pyshv as either.

The one non-standard error we keep is UNKNOWN that doesn't have its
class defined because RpcError is used for it. That is implementations
defined error which is allowed but should never be sent to SHV.