rpctransport/tcp: added support for SSL
Thanks to the Python support the SSL support is pretty much included in
the standard TCP/IP. The only difference is that we have to be able to
provide correct SSL context.

This doesn't include any integration to the Broker. We could in the
future use client certificates to ease the client login (skip the
requirement for the password).