Fix after upstream changes
Upstream removed access to the pythonPackages for python package
functions. The alternative is to use python.pkgs.

This also modifies the overlay names. This makes them more consistent
and explicit. pythonPackages is shorter with same meaning as previous
one. pkgs is better beacuse it is overlay for the nixpkgs package set.
The original noInherit wan't clear what it should be applied on.