rpcmessage: changes in RpcMessage for Response Delay
This introduces message extensions and changes that were introduced in
SHV standard (shv-doc repository) commit
26b7272a73610fb8055561c2075d069e56fd9f57.

This also makes some of the message parameters access more consistent
and more strict. In the past it was implemented to do the best effort.
This moves the implementation rather to the canonical standard
correctness.