rpcmethod: rename RpcMethodDesc to RpcDir and split off RpcAccess
This refactors the `dir` related API. The access representation is split
to its separate definition. This is mostly only technical change in the
API. It will require replacement of the names in the code but nothing
more.

The old API is not aliased due to this project being still in unstable
phase and thus API can be easilly modified in this way.