rpctransport/stream: prevent possible message spliting
This ensures that even if multiple sends are performed in parallel that
single message is submitted as one complete sequence. With previous
implementation it was possible that TTY writes would be mixed up. On
asyncio streams we were probably saved that only actual await that might
switch to different task was drain itself and thus no splitting would be
possible.

The new approach is technically correct but at the same time we did not
see any wrong behavior with old implementation.