shvbase: handle request aborts
This adds support for request abort messages. Now they are handled and
progress is reported with response delay message.

This primary missing part right now is in the form of self response
delay sending. This should be implemented in the way that these kind of
messages are sent only when there is bandwidth for them and we have no
instrumentation for that right now.

The secondary change is the provided access to the caller_ids in
requests to allow peer specific contexts.