shvbase: Implement _idle_message
This solves an issue with bounded bandwidth. It is likely that we can
have more messages than we can pass through the limited bandwidth. In
SHV there are messages such as signals or some of the delayed responses
we can send only if there is bandwidth for them.

The implementation introduces central loop for the message sending. This
way we can detect that there are no more messages to send and ask for
some additional messages to be created.

This infrastructure is immediately used to implement opportunistic
sending of the delayed response progress. It also fit pretty well to the
broker and thus Broker now uses this as well to propagate messages from
other peers.