rpctransport/pipe: fix error when using on Python 3.12 and newer
The Python 3.12 change the behavior of the StreamReaderProtocol. Now the
error is emited if StreamReaderProtocol is made ready two times which is
what happens when we use connect_write_pipe. This wasn't the case before
3.12. The new approach that seems to work is to have dedicated protocol
for writing by passing `None` to the reader but that is not allowed
before 3.12 and thus we just have two different implementations in the
code at the moment.