cpon: fix parsing of float numbers
The previous version wasn't able to parse decimal value provided as
mantisa. The naive implementation where for both variants is used
`float.fromhex` seems to work for numbers greater than 1 but not for
those between 0 and 1. This this alternative implementation is provided.

The invalid behavior was introduced in
a939b5d09e3a17484e7264d378722469346f4282.