rpctransport/ws: use subprotocol to filter out other traffic
Other implementations of SHV RPC have historical version of websockets
that are based on just fragmented block stream protocol. To ensure that
we don't collide with them we now require `shv3` to be provided as
subprotocol. The connection won't be established if peer fails to do so.

The old pre SHV3 websockets are not planned to be implemented in pySHV.