gitlab-ci: fix release creation
The pip fails here as it is no longer allowed to install python packages
to the system by pip. But we don't need these packages. They were pulled
in because they are required for C release.sh script that manages
resources but Python projects do not upload any release resources and
thus there is no need for these tools.