commonpack: optimize unpack by removing read_raw and read_meta
The methods read_raw is not used in the code and it only complicates
logic. The new logic is much more streamlined in therms of peeked byte
usage and also increases overal performance of the code as _read is very
heavily called method.

read_meta is also not used and it makes sense to reformat read methods
to return directly instead of what we had so far.