rpctransport/tty: remove dependency on pyserial
The pyserial implementation seems to be problematic on Unix platforms.
We can use builtin asyncio facilities to get the same functionality with
ease and be even more compatible with rest of the code. On the other
hand we can't support Windows this way. The Windows support was broken
since 349d1dd54b5009112d4c3923d315d970341a242c. The support can be later
reintroduced by again pulling in the pyserial dependency but only for
the Windows platform.