shvbase: cancel loop on disconnect
This ensures that loop is terminated on the cancel to prevent situations
when reader is not correctly tied into the writer and thus receive in
the loop would not terminate. The effect would be stuck client.

Sockets as they are provided by asyncio are capable of this propagation
but we don't test this for clients and in general testing for it will
just generate unhandled errors in asyncio. Thus it seems that the
correct solution is to terminate explicitly rather than implicitly.