rpclogin: do not force SHA1 if on secure transport
The newly added SSL ensures that we can send login in plain text. This
is required if broker will use the password for login to some third
party login provider that doesn't support the SHV's SHA1 scheme. To
ensure that password is not disclosed except to the two participants we
require the transport to be encrypted.