shvclient: allow sending other login messages
The login might be a bit more complex over just standard one that is
provided. Thus we have to allow not only known methods but all methods
on the root node. We do not allow only usage of ls and dir as those are
not defined for the login phase and instead are expected to be the
possible initial call right after successful login.