shvclient: allow extension of _login and _post_login
It is better to split existing _login task to two phases where first we
perform login and then we perform post-login tasks. This provides users
of this class a way to extend both of the parts independently.

The login is now expected to be customized to support non-standard login
schemes. Only the standard login schemes are available in pyshv.