treewide: unexport most of the interface from top level package
The inclusion of all types to the top level wasn't a good practice for
this project. The issue was that import of `shv` package caused load of
all components (except few) which was consuming considerable amount of
time. This is negligible on normal PC but on embedded Linux it turns out
it can be seconds. By removing import in the root package and requiring
applications to import only selected part the overall application
startup should be shorter.

The test of import time before and after these changes:
  import time:       360 |     226844 | shv.rpcapi.client
And before:
  import time:       745 |     386800 | shv

This commit doesn't contain any API change. It includes only changes in
imports. The short list for the important classes that are commonly used
by external projects and their suggested import location:

- from shv.rpcapi import SHVBase
- from shv.rpcapi.client import SHVClient
- from shv.rpcapi.valueclient import SHVValueClient
- from shv.rpcapi.methods import SHVMethods
- from shv.rpcdef import RpcAccess, RpcDir, RpcError
- from shv.rpcdef.alert import RpcAlert
- from shv.rpcdef.file import RpcFleStat, RpcFile, FileProvider
- from shv.path import SHVPath, SHVPathParents
- from shv.rpcurl import RpcUrl, RpcProtocol
- from shv.rpclogin import RpcLogin, RpcLoginType
- from shv.rpcmessage import RpcMessage
- from shv.rpcri import rpcri_match, rpcri_relative_to
- from shv.rpcparam import shvarg, shvargt, shvget, shvgett, shvt