commonpack: do not use EOFError
This removes usage of EOFError in unpackers. The reason is that it is
confusing from the user point of view. It actually doesn't matter if
issue is either due to missing or invalid bytes. The point is that the
encoding is invalid. The exact reason is stated in the message. By
requiring users to catch both ValueError and EOFError we are asking for
issues when they inevitably forget to catch one of them (which is how we
discovered this).