rpcapi/client: use cancel to restart login explicitly
Login task is being cancelled in case of restart. This raises
CancelledError from awaits for that task. To cover this we suppress this
error but at the same time that removes ability to timeout this method
as it will be caught in the infinite loop.

The solution here is to just signal with message that cancellation is
due to login task restart and ignore just that.