rpcri: add support for ** at tail matching prefix
The change in the SHV 3.0 standard introduced change in the handling of
the `**` pattern. It is now given that if it is at the end of the path
it will also match prefix without any additional node. Thus `foo/**`
matches not only `foo/bar` and so on but also `foo` itself now.

The reference to the shv-doc commit:
  826b304dc91ee6a82b038b9e2b18be25a15d7f49